[p4runtime session] Pull test helper functions into library so we can reuse them.
